+++ libs/syntax/CHANGELOG.md
@@ -3,12 +3,12 @@
 ## [Unreleased]
 
 ### Added
-- Handle `important` field property.
+- Handle `important`, `restricted` and `ownership` field property.
 - Handle `underscoreUpper` for names (Example: FIRST_NAME).
 - Handle `hyphenUpper` for names (Example: FIRST-NAME).
 - Handle subtype `rich` of type `string`.
 
 ### Changed
-- Use of `safe-eval` for evalutation
+- Use of `safe-eval` for evaluation
 
 ### Removed